prepare_projects: only init subrepos, if no specific project is provided
The fix in f2142f82b ("prepare_projects: only checkout real subrepos")
was not enough: If a repo subdir in a project configuration contains a
non-gitmodules managed repository, prepare_project will errorneously
detect it as a subrepo and try to check it out.

This patch prevents this behaviour. Only update submodules, if no
specific project name is provided.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>